WIP: finish configureStore integration:
